third_party/blink/web_tests/http/tests/devtools/elements/edit/set-outer-html-2-expected.txt

Tests DOMAgent.setOuterHTML protocol method (part 2).


Running: testSetUp

Running: testChangeMultipleThings
Wrapper identity: identity
Event AttrModified: H2
Event AttrModified: H2
Event NodeInserted: UL
Event NodeRemoved: UL
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2 foo="bar" bar="baz">Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      <ul>
         
      </ul>
      <ul>
         <li></li>
      </ul>
      </div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event AttrRemoved: H2
Event AttrRemoved: H2
Event NodeInserted: LI
Event NodeRemoved: #text
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      <ul>
         <li></li>
      </ul>
      <ul>
         <li></li>
      </ul>
      </div>
==========>8==========





Running: testChangeNestingLevel
Wrapper identity: identity
Event NodeInserted: DIV
Event NodeRemoved: UL
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      <div><ul>
         <li></li>
      </ul></div>
      <ul>
         <li></li>
      </ul>
      </div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event NodeInserted: UL
Event NodeRemoved: DIV
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      <ul>
         <li></li>
      </ul>
      <ul>
         <li></li>
      </ul>
      </div>
==========>8==========





Running: testSwapNodes
Wrapper identity: identity
Event NodeInserted: H2
Event NodeRemoved: H2
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      
      <p id="identity">There are many ways to get involved. You can:</p>
      <ul>
         <li></li>
      </ul>
      <ul>
         <li></li>
      </ul>
      <h2>Getting involved</h2></div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event NodeInserted: H2
Event NodeInserted: P
Event NodeInserted: UL
Event NodeInserted: UL
Event NodeRemoved: H2
Event NodeRemoved: P
Event NodeRemoved: UL
Event NodeRemoved: UL
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      <ul>
         <li></li>
      </ul>
      <ul>
         <li></li>
      </ul>
      </div>
==========>8==========





Running: testEditTwoRoots
Wrapper identity: identity
Event NodeInserted: DIV
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      <ul>
         <li></li>
      </ul>
      <ul>
         <li></li>
      </ul>
      </div>
==========>8==========

Bringing things back

Wrapper identity: identity
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      <ul>
         <li></li>
      </ul>
      <ul>
         <li></li>
      </ul>
      </div>
==========>8==========





Running: testDupeNode
Replacing '<h2>Getting involved</h2>' with '<h2>Getting involved</h2><h2>Getting involved</h2>'

Wrapper identity: identity
Event NodeInserted: H2
Event NodeInserted: H2
Event NodeRemoved: H2
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2><h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      <ul>
         <li></li>
      </ul>
      <ul>
         <li></li>
      </ul>
      </div>
==========>8==========

Bringing things back

Wrapper identity: identity
Event NodeRemoved: H2
==========8<==========
<div id="container" style="display:none">
      <p>WebKit is used by <a href="http://www.apple.com/safari/">Safari</a>, Dashboard, etc..</p>
      <h2>Getting involved</h2>
      <p id="identity">There are many ways to get involved. You can:</p>
      <ul>
         <li></li>
      </ul>
      <ul>
         <li></li>
      </ul>
      </div>
==========>8==========
